fixed bug for self identification module and reading gas and smoke sensor

+++ src/libs/MqttHelper/MqttHelper.h
@@ -10,7 +10,7 @@ MqttHelper::MqttHelper(Config *config = new Config(), Sensor *relay = nullptr, i
         Serial.println("IP");
         int Parts[4] = {0, 0, 0, 0};
         int Part = 0;
-        for (int i = 0; i < address.length(); i++)
+        for (size_t i = 0; i < address.length(); i++)
         {
             char c = address[i];
             if (c == '.')
